fix(sysid): give the fim column index an explicit integer dtype
CI's typecheck job failed tier 1 with two pyright errors at
`names = tuple(names[i] for i in idx)`: under the numpy 2.5.3 stubs CI
resolves (this box has 2.4.6), `np.asarray(<list of int>)` is typed
float64, so indexing a tuple with its elements is an overload error.
Reproduced exactly with pyright 1.1.414 against numpy 2.5.3 + jax 0.10.2
in a scratch --target; 0 errors against the local environment, which is why
no local run saw it.  `dtype=np.intp` is what np.asarray already produced at
runtime on 64-bit Linux, so behaviour is bit-identical: 230 fim tests pass,
golden record and transfer counts included.

+++ src/maddening/sysid.py
@@ -621,7 +621,7 @@ def _masked_indices(params: dict, mask: Optional[dict]) -> Optional[np.ndarray]:
         offset += n
     if not idx:
         raise ValueError("mask selects no parameters")
-    return np.asarray(idx)
+    return np.asarray(idx, dtype=np.intp)
 
 
 def _fim_indices(params: dict, mask: Optional[dict]):
@@ -684,7 +684,7 @@ def _fim_indices(params: dict, mask: Optional[dict]):
         offset += n
     if not idx:
         raise ValueError("mask selects no parameters")
-    return np.asarray(idx), ()
+    return np.asarray(idx, dtype=np.intp), ()
 
 
 def _refuse_integer_trainable(params: dict, mask: dict) -> None:
